+++ npm_and_yarn/spec/dependabot/npm_and_yarn/file_updater_spec.rb
@@ -1683,528 +1683,7 @@ RSpec.describe Dependabot::NpmAndYarn::FileUpdater do
       end
     end
 
-    ############################################################
-    # Tests for error cases. Must test npm and Yarn separately #
-    ############################################################
-    describe "errors" do
-      context "with a dependency version that can't be found" do
-        let(:manifest_fixture_name) { "yanked_version.json" }
-        let(:npm_lock_fixture_name) { "yanked_version.json" }
-        let(:yarn_lock_fixture_name) { "yanked_version.lock" }
-
-        context "with a yarn lockfile" do
-          let(:files) { [package_json, yarn_lock] }
-          it "raises a helpful error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::DependencyFileNotResolvable)
-          end
-        end
-      end
-
-      context "with a sub dependency name that can't be found" do
-        let(:manifest_fixture_name) do
-          "github_sub_dependency_name_missing.json"
-        end
-        let(:npm_lock_fixture_name) do
-          "github_sub_dependency_name_missing.json"
-        end
-        let(:yarn_lock_fixture_name) do
-          "github_sub_dependency_name_missing.lock"
-        end
-
-        let(:dependency_name) { "test-missing-dep-name-npm-package" }
-        let(:requirements) do
-          [{
-            requirement: nil,
-            file: "package.json",
-            groups: ["dependencies"],
-            source: {
-              type: "git",
-              url: "https://github.com/dependabot-fixtures/"\
-                   "test-missing-dep-name-npm-package",
-              branch: nil,
-              ref: ref
-            }
-          }]
-        end
-        let(:previous_requirements) do
-          [{
-            requirement: nil,
-            file: "package.json",
-            groups: ["dependencies"],
-            source: {
-              type: "git",
-              url: "https://github.com/dependabot-fixtures/"\
-                   "test-missing-dep-name-npm-package",
-              branch: nil,
-              ref: old_ref
-            }
-          }]
-        end
-        let(:previous_version) { "1be88e036981a8511eacf3f20e0a21507349988d" }
-        let(:version) { "346e79a12f34c5937bf0f016bced0723f864fe19" }
-        let(:ref) { "v1.0.1" }
-        let(:old_ref) { "v1.0.0" }
-
-        context "with an npm lockfile" do
-          let(:files) { [package_json, package_lock] }
-
-          it "raises a helpful error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::DependencyFileNotResolvable)
-          end
-        end
-
-        context "with a yarn lockfile" do
-          let(:files) { [package_json, yarn_lock] }
-
-          it "raises a helpful error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::DependencyFileNotResolvable)
-          end
-        end
-      end
-
-      context "with an invalid requirement in the package.json" do
-        let(:manifest_fixture_name) { "invalid_requirement.json" }
-        let(:npm_lock_fixture_name) { "package-lock.json" }
-        let(:yarn_lock_fixture_name) { "yarn.lock" }
-
-        context "with an npm lockfile" do
-          let(:files) { [package_json, package_lock] }
-
-          it "raises a helpful error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::DependencyFileNotResolvable)
-          end
-        end
-
-        context "with a yarn lockfile" do
-          let(:files) { [package_json, yarn_lock] }
-
-          it "raises a helpful error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::DependencyFileNotResolvable)
-          end
-        end
-      end
-
-      context "when there is a private dep we don't have access to" do
-        let(:manifest_fixture_name) { "private_source.json" }
-        let(:npm_lock_fixture_name) { "private_source.json" }
-        let(:yarn_lock_fixture_name) { "private_source.lock" }
-
-        context "with a yarn lockfile" do
-          let(:files) { [package_json, yarn_lock] }
-
-          it "raises a helpful error" do
-            expect { updater.updated_dependency_files }.
-              to raise_error(Dependabot::PrivateSourceAuthenticationFailure)
-          end
-
-          context "that is unscoped" do
-            let(:manifest_fixture_name) { "private_source_unscoped.json" }
-            let(:yarn_lock_fixture_name) { "private_source_unscoped.lock" }
-
-            let(:dependency_name) { "my-private-dep-asdlfkasdf" }
-            let(:version) { "1.0.1" }
-            let(:previous_version) { "1.0.0" }
-            let(:requirements) do
-              [{
-                file: "package.json",
-                requirement: "^1.0.0",
-                groups: ["devDependencies"],
-                source: nil
-              }]
-            end
-            let(:previous_requirements) do
-              [{
-                file: "package.json",
-                requirement: "^1.0.0",
-                groups: ["devDependencies"],
-                source: nil
-              }]
-            end
-
-            it "raises a helpful error" do
-              expect { updater.updated_dependency_files }.
-                to raise_error do |error|
-                  expect(error).
-                    to be_a(Dependabot::PrivateSourceAuthenticationFailure)
-                  expect(error.source).to eq("npm-proxy.fury.io/<redacted>")
-                end
-            end
-
-            context "with bad credentials" do
-              let(:credentials) do
-                [{
-                  "type" => "git_source",
-                  "host" => "github.com",
-                  "username" => "x-access-token",
-                  "password" => "token"
-                }, {
-                  "type" => "npm_registry",
-                  "registry" => "npm-proxy.fury.io/dependabot",
-                  "token" => "bad_token"
-                }]
-              end
-
-              # TODO: Fix broken test
-              it "raises a helpful error" do
-                expect { updater.updated_dependency_files }.
-                  to raise_error(Dependabot::PrivateSourceAuthenticationFailure)
-              end
-            end
-          end
-        end
-      end
-
-      context "when there is a private git dep we don't have access to" do
-        let(:manifest_fixture_name) { "github_dependency_private.json" }
-        let(:npm_lock_fixture_name) { "github_dependency_private.json" }
-        let(:yarn_lock_fixture_name) { "github_dependency_private.lock" }
-
-        let(:dependency_name) { "strict-uri-encode" }
-        let(:version) { "1.1.0" }
-        let(:requirements) { [] }
-
-        context "with a yarn lockfile" do
-          let(:files) { [package_json, yarn_lock] }
-
-          it "raises a helpful error" do
-            expect { updater.updated_dependency_files }.
-              to raise_error(Dependabot::GitDependenciesNotReachable) do |error|
-                expect(error.dependency_urls).
-                  to eq(
-                    [
-                      "ssh://git@github.com/hmarr/"\
-                      "dependabot-test-private-npm-package.git"
-                    ]
-                  )
-              end
-          end
-        end
-
-        context "with a npm lockfile" do
-          let(:files) { [package_json, package_lock] }
-
-          it "raises a helpful error" do
-            expect { updater.updated_dependency_files }.
-              to raise_error(Dependabot::GitDependenciesNotReachable) do |error|
-                expect(error.dependency_urls).
-                  to eq(
-                    [
-                      "ssh://git@github.com/hmarr/"\
-                      "dependabot-test-private-npm-package.git"
-                    ]
-                  )
-              end
-          end
-        end
-      end
-
-      context "because we're updating to a nonexistent version" do
-        let(:yarn_lock_fixture_name) { "yarn.lock" }
-        let(:npm_lock_fixture_name) { "package-lock.json" }
-        let(:manifest_fixture_name) { "package.json" }
-
-        let(:dependency_name) { "fetch-factory" }
-        let(:version) { "5.0.2" }
-        let(:requirements) do
-          [{
-            file: "package.json",
-            requirement: "^5.0.2",
-            groups: ["dependencies"],
-            source: nil
-          }]
-        end
-
-        it "raises an unhandled error" do
-          expect { updated_files }.
-            to raise_error(Dependabot::InconsistentRegistryResponse)
-        end
-      end
-
-      context "with a dependency that can't be found" do
-        let(:manifest_fixture_name) { "nonexistent_dependency.json" }
-        let(:npm_lock_fixture_name) { "yanked_version.json" }
-        let(:yarn_lock_fixture_name) { "yanked_version.lock" }
-
-        context "with an npm lockfile" do
-          let(:files) { [package_json, package_lock] }
-
-          it "raises a helpful error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::PrivateSourceAuthenticationFailure)
-          end
-        end
-
-        context "with a yarn lockfile" do
-          let(:files) { [package_json, yarn_lock] }
-
-          it "raises a helpful error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::PrivateSourceAuthenticationFailure)
-          end
-        end
-      end
-
-      context "with a registry that times out" do
-        let(:manifest_fixture_name) { "package.json" }
-        let(:yarn_lock_fixture_name) { "yarn.lock" }
-        let(:yarnrc) do
-          Dependabot::DependencyFile.new(
-            name: ".yarnrc",
-            content: 'registry "https://timeout.cv/repository/mirror/"'
-          )
-        end
-
-        context "with a yarn lockfile" do
-          let(:files) { [package_json, yarn_lock, yarnrc] }
-          # This test is extremely slow (it take 1m45 to run) so should only be
-          # run locally.
-          # it "raises a helpful error" do
-          #   expect { updated_files }.
-          #     to raise_error(Dependabot::PrivateSourceTimedOut) do |error|
-          #       expect(error.source).to eq("timeout.cv/repository/mirror")
-          #     end
-          # end
-        end
-      end
-
-      context "with a git reference that Yarn would find but npm wouldn't" do
-        let(:manifest_fixture_name) { "git_dependency_yarn_ref.json" }
-        let(:npm_lock_fixture_name) { "git_dependency_yarn_ref.json" }
-
-        context "with an npm lockfile" do
-          let(:files) { [package_json, package_lock] }
-          it "raises a helpful error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::DependencyFileNotResolvable)
-          end
-        end
-      end
-
-      context "with a corrupted npm lockfile (version missing)" do
-        let(:manifest_fixture_name) { "package.json" }
-        let(:npm_lock_fixture_name) { "version_missing.json" }
-
-        context "with an npm lockfile" do
-          let(:files) { [package_json, package_lock] }
-          it "raises a helpful error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::DependencyFileNotResolvable) do |error|
-                expect(error.message).
-                  to include(
-                    "lockfile has some corrupt entries with missing versions"
-                  )
-              end
-          end
-        end
-      end
-
-      context "when scoped sub dependency version is missing" do
-        let(:manifest_fixture_name) do
-          "github_scoped_sub_dependency_version_missing.json"
-        end
-        let(:npm_lock_fixture_name) do
-          "github_scoped_sub_dependency_version_missing.json"
-        end
-        let(:yarn_lock_fixture_name) do
-          "github_scoped_sub_dependency_version_missing.lock"
-        end
-
-        let(:dependency_name) do
-          "@dependabot/test-missing-scoped-dep-version-npm-package"
-        end
-        let(:requirements) do
-          [{
-            requirement: nil,
-            file: "package.json",
-            groups: ["dependencies"],
-            source: {
-              type: "git",
-              url: "https://github.com/dependabot-fixtures/"\
-                   "test-missing-scoped-dep-version-npm-package",
-              branch: nil,
-              ref: ref
-            }
-          }]
-        end
-        let(:previous_requirements) do
-          [{
-            requirement: nil,
-            file: "package.json",
-            groups: ["dependencies"],
-            source: {
-              type: "git",
-              url: "https://github.com/dependabot-fixtures/"\
-                   "test-missing-scoped-dep-version-npm-package",
-              branch: nil,
-              ref: old_ref
-            }
-          }]
-        end
-        let(:previous_version) { "fe5138f33735fb07891d348cd1e985fe3134211c" }
-        let(:version) { "7abd161c8eba336f06173f0a97bc8decc3cd9c2c" }
-        let(:ref) { "v1.0.4" }
-        let(:old_ref) { "v1.0.3" }
-
-        context "npm only" do
-          let(:files) { [package_json, package_lock] }
-
-          it "raises an error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::InconsistentRegistryResponse)
-          end
-        end
-
-        context "yarn only" do
-          let(:files) { [package_json, yarn_lock] }
-
-          it "raises an error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::InconsistentRegistryResponse)
-          end
-        end
-      end
-
-      context "when sub dependency version is missing" do
-        let(:manifest_fixture_name) do
-          "github_sub_dependency_version_missing.json"
-        end
-        let(:npm_lock_fixture_name) do
-          "github_sub_dependency_version_missing.json"
-        end
-        let(:yarn_lock_fixture_name) do
-          "github_sub_dependency_version_missing.lock"
-        end
-
-        let(:dependency_name) { "test-missing-dep-version-npm-package" }
-        let(:requirements) do
-          [{
-            requirement: nil,
-            file: "package.json",
-            groups: ["dependencies"],
-            source: {
-              type: "git",
-              url: "https://github.com/dependabot-fixtures/"\
-                   "test-missing-dep-version-npm-package",
-              branch: nil,
-              ref: ref
-            }
-          }]
-        end
-        let(:previous_requirements) do
-          [{
-            requirement: nil,
-            file: "package.json",
-            groups: ["dependencies"],
-            source: {
-              type: "git",
-              url: "https://github.com/dependabot-fixtures/"\
-                   "test-missing-dep-version-npm-package",
-              branch: nil,
-              ref: old_ref
-            }
-          }]
-        end
-        let(:previous_version) { "f56186c1643a9a09a86dfe09b1890921330c28bb" }
-        let(:version) { "3deb2768be1591f2fdbdec40aa6a63ba6b270b40" }
-        let(:ref) { "v1.0.3" }
-        let(:old_ref) { "v1.0.2" }
-
-        context "npm only" do
-          let(:files) { [package_json, package_lock] }
-
-          it "raises an error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::SharedHelpers::HelperSubprocessFailed)
-          end
-        end
-
-        context "yarn only" do
-          let(:files) { [package_json, yarn_lock] }
-
-          it "raises an error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::SharedHelpers::HelperSubprocessFailed)
-          end
-        end
-      end
-
-      context "when a git src dependency doesn't have a valid package.json" do
-        let(:manifest_fixture_name) { "git_missing_version.json" }
-        let(:npm_lock_fixture_name) { "git_missing_version.json" }
-        let(:yarn_lock_fixture_name) { "git_missing_version.lock" }
-
-        let(:dependency_name) { "raven-js" }
-        let(:requirements) do
-          [{
-            requirement: nil,
-            file: "package.json",
-            groups: ["dependencies"],
-            source: {
-              type: "git",
-              url: "https://github.com/getsentry/raven-js",
-              branch: nil,
-              ref: ref
-            }
-          }]
-        end
-        let(:previous_requirements) do
-          [{
-            requirement: nil,
-            file: "package.json",
-            groups: ["dependencies"],
-            source: {
-              type: "git",
-              url: "https://github.com/getsentry/raven-js",
-              branch: nil,
-              ref: old_ref
-            }
-          }]
-        end
-        let(:previous_version) { "c2b377e7a254264fd4a1fe328e4e3cfc9e245570" }
-        let(:version) { "70b24ed25b73cc15472b2bd1c6032e22bf20d112" }
-        let(:ref) { "4.4.1" }
-        let(:old_ref) { "3.23.1" }
-
-        context "npm only" do
-          let(:files) { [package_json, package_lock] }
-
-          it "raises a HelperSubprocessFailed error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::DependencyFileNotResolvable)
-          end
-        end
-
-        context "yarn only" do
-          let(:files) { [package_json, yarn_lock] }
-
-          it "raises helpful error" do
-            expect { updated_files }.
-              to raise_error(Dependabot::DependencyFileNotResolvable)
-          end
-        end
-      end
-
-      context "with an invalid package name" do
-        let(:manifest_fixture_name) { "invalid_package_name.json" }
-        let(:npm_lock_fixture_name) { "invalid_package_name.json" }
-        let(:dependency_name) { "fetch-factory:" }
-
-        # Currently only npm returns a useful error
-        context "with an npm lockfile" do
-          let(:files) { [package_json, package_lock] }
-
-          it "raises a helpful error" do
-            expect { updater.updated_dependency_files }.
-              to raise_error(Dependabot::DependencyFileNotResolvable)
-          end
-        end
-      end
-
+    context "errors" do
       context "git sub-dependency with invalid from" do
         let(:manifest_fixture_name) { "git_sub_dep_invalid_from.json" }
         let(:npm_lock_fixture_name) { "git_sub_dep_invalid_from.json" }
